+++ editor/code/src/config.ts
@@ -8,10 +8,14 @@ export interface CoqLspServerConfig {
   show_coq_info_messages: boolean;
   show_notices_as_diagnostics: boolean;
   admit_on_bad_qed: boolean;
+  debug: boolean;
 }
 
 export namespace CoqLspServerConfig {
-  export function create(client_version: string, wsConfig: any) {
+  export function create(
+    client_version: string,
+    wsConfig: any
+  ): CoqLspServerConfig {
     return {
       client_version: client_version,
       eager_diagnostics: wsConfig.eager_diagnostics,
@@ -20,6 +24,7 @@ export namespace CoqLspServerConfig {
       show_coq_info_messages: wsConfig.show_coq_info_messages,
       show_notices_as_diagnostics: wsConfig.show_notices_as_diagnostics,
       admit_on_bad_qed: wsConfig.admit_on_bad_qed,
+      debug: wsConfig.debug,
     };
   }
 }