+++ fleche/config.ml
@@ -21,6 +21,8 @@ type t =
             when a [Qed] fails: one is not to add the constant to the state, the
             other one is admit it. We find the second behavior more useful, but
             YMMV. *)
+  ; debug : bool [@default false]
+        (** Enable debug on Coq side, including backtraces *)
   }
 
 let default =
@@ -33,6 +35,7 @@ let default =
   ; show_coq_info_messages = false
   ; show_notices_as_diagnostics = false
   ; admit_on_bad_qed = true
+  ; debug = false
   }
 
 let v = ref default